NP-3537 Remove outdated edit fucntionality for old affiliation IDs
