--- src/pages/user/institution/InstitutionCard.tsx
@@ -1,98 +1,0 @@
-import { useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useDispatch, useSelector } from 'react-redux';
-import { Box, Button, Paper } from '@mui/material';
-import DeleteIcon from '@mui/icons-material/Delete';
-import EditIcon from '@mui/icons-material/Edit';
-import { AuthorityQualifiers, updateQualifierIdForAuthority } from '../../../api/authorityApi';
-import { AffiliationHierarchy } from '../../../components/institution/AffiliationHierarchy';
-import { EditInstitution } from '../../../components/institution/EditInstitution';
-import { StyledRightAlignedWrapper } from '../../../components/styled/Wrappers';
-import { setNotification } from '../../../redux/actions/notificationActions';
-import { setAuthorityData } from '../../../redux/actions/userActions';
-import { RootStore } from '../../../redux/reducers/rootReducer';
-import { FormikInstitutionUnit } from '../../../types/institution.types';
-import { getMostSpecificUnit } from '../../../utils/institutions-helpers';
-import { cristinBaseId, isErrorStatus, isSuccessStatus } from '../../../utils/constants';
-
-interface InstitutionCardProps {
-  orgunitId: string;
-  setInstitutionIdToRemove: (orgunitId: string) => void;
-}
-
-export const InstitutionCard = ({ orgunitId, setInstitutionIdToRemove }: InstitutionCardProps) => {
-  const { t } = useTranslation('common');
-  const [openEditForm, setOpenEditForm] = useState(false);
-  const dispatch = useDispatch();
-  const { user } = useSelector((store: RootStore) => store);
-
-  // TODO: Remove when all IDs are updated (NP-3492)
-  const handleEditInstitution = async (values: FormikInstitutionUnit, initialInstitution: string) => {
-    if (!values.unit || !user?.authority) {
-      return;
-    }
-
-    const mostSpecificUnit = getMostSpecificUnit(values.unit);
-    const newUnitId = mostSpecificUnit.id;
-
-    if (!newUnitId) {
-      return;
-    } else if (user.authority.orgunitids.includes(newUnitId)) {
-      dispatch(setNotification(t('feedback:info.affiliation_already_exists'), 'info'));
-      return;
-    }
-
-    const updateAuthorityResponse = await updateQualifierIdForAuthority(
-      user.authority.id,
-      AuthorityQualifiers.OrgUnitId,
-      initialInstitution,
-      newUnitId
-    );
-    if (isErrorStatus(updateAuthorityResponse.status)) {
-      dispatch(
-        setNotification(
-          t('feedback:error.update_authority', { qualifier: t(`common:${AuthorityQualifiers.OrgUnitId}`) }),
-          'error'
-        )
-      );
-    } else if (isSuccessStatus(updateAuthorityResponse.status)) {
-      dispatch(setAuthorityData(updateAuthorityResponse.data));
-      dispatch(setNotification(t('feedback:success.added_affiliation')));
-    }
-  };
-
-  return openEditForm ? (
-    <Paper sx={{ p: '1rem' }}>
-      <EditInstitution
-        initialInstitutionId={orgunitId}
-        onSubmit={(values) => handleEditInstitution(values, orgunitId)}
-        onCancel={() => setOpenEditForm(false)}
-      />
-    </Paper>
-  ) : (
-    <Paper elevation={3} data-testid="institution-presentation" sx={{ p: '1rem' }}>
-      <Box>
-        <AffiliationHierarchy unitUri={orgunitId} />
-      </Box>
-      <StyledRightAlignedWrapper sx={{ mt: '0.5rem' }}>
-        {orgunitId.includes(cristinBaseId) && (
-          <Button
-            variant="outlined"
-            data-testid={`button-edit-institution-${orgunitId}`}
-            startIcon={<EditIcon />}
-            onClick={() => setOpenEditForm(true)}>
-            {t('edit')}
-          </Button>
-        )}
-        <Button
-          color="error"
-          variant="outlined"
-          data-testid={`button-delete-institution-${orgunitId}`}
-          startIcon={<DeleteIcon />}
-          onClick={() => setInstitutionIdToRemove(orgunitId)}>
-          {t('remove')}
-        </Button>
-      </StyledRightAlignedWrapper>
-    </Paper>
-  );
-};
