+++ src/types/institution.types.ts
@@ -1,6 +1,6 @@
 import { LanguageString } from './common.types';
 
-export interface InstitutionUnitBase {
+interface InstitutionUnitBase {
   acronym?: string;
   name: string;
   id: string;
@@ -10,19 +10,9 @@ export interface RecursiveInstitutionUnit extends InstitutionUnitBase {
   subunits?: RecursiveInstitutionUnit[];
 }
 
-export interface FormikInstitutionUnit extends Partial<RecursiveInstitutionUnit> {
-  subunit?: InstitutionUnitBase;
-  unit?: RecursiveInstitutionUnit;
-}
-
 export enum FormikInstitutionUnitFieldNames {
   SubUnit = 'subunit',
   Unit = 'unit',
-}
-
-export interface InstitutionState {
-  items: InstitutionUnitBase[];
-  language?: string;
 }
 
 export interface Organization {
