+++ src/utils/institutions-helpers.ts
@@ -1,19 +1,6 @@
 import { Contributor } from '../types/contributor.types';
-import {
-  FormikInstitutionUnit,
-  InstitutionUnitBase,
-  Organization,
-  RecursiveInstitutionUnit,
-} from '../types/institution.types';
+import { Organization, RecursiveInstitutionUnit } from '../types/institution.types';
 import { getLanguageString } from './translation-helpers';
-
-// Find the most specific unit in hierarchy
-export const getMostSpecificUnit = (values: FormikInstitutionUnit): InstitutionUnitBase => {
-  if (values.subunit) {
-    return getMostSpecificUnit(values.subunit);
-  }
-  return values as InstitutionUnitBase;
-};
 
 // Find distinct unit URIs for a set of contributors' affiliations
 const unitIdToIgnore = 'https://api.cristin.no/v2/units/0.0.0.0';
@@ -69,7 +56,7 @@ const getAllChildOrganizations = (units: Organization[] = [], result: Organizati
 
 // converts from https://api.cristin.no/v2/units/7482.3.3.0
 //            to https://api.cristin.no/v2/institutions/7482
-export const convertToInstitution = (unitId: string) => {
+const convertToInstitution = (unitId: string) => {
   if (unitId.includes('/institutions/')) {
     return unitId;
   } else {
@@ -78,12 +65,3 @@ export const convertToInstitution = (unitId: string) => {
     return `https://api.cristin.no/v2/institutions/${institutionId}`;
   }
 };
-
-export const sortInstitutionsAlphabetically = (institutions: InstitutionUnitBase[]) =>
-  institutions.sort((institution1, institution2) => {
-    if (institution1.name.toLocaleLowerCase() < institution2.name.toLocaleLowerCase()) {
-      return -1;
-    } else {
-      return 1;
-    }
-  });
